+++ styles/indexStyles.tsx
@@ -1,0 +1,15 @@
+import React, { useState } from "react";
+import {
+  Text,
+  View,
+  Button,
+  Modal,
+  TextInput,
+  StyleSheet,
+  FlatList,
+  Dimensions,
+  Pressable,
+} from "react-native";
+
+
+export default styles;
