--- styles/styles.tsx
@@ -1,165 +1,0 @@
-import React, { useState } from "react";
-import {
-  Text,
-  View,
-  Button,
-  Modal,
-  TextInput,
-  StyleSheet,
-  FlatList,
-  Dimensions,
-  Pressable,
-} from "react-native";
-
-const { width } = Dimensions.get("window");
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-  },
-  modalView: {
-    width: "70%",
-    backgroundColor: "white",
-    borderRadius: 10,
-    padding: 20,
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    elevation: 5,
-  },
-  modalText: {
-    fontSize: 18,
-    marginBottom: 15,
-  },
-  eventNameInput: {
-    width: "70%",
-    height: 40,
-    borderColor: "gray",
-    borderWidth: 1,
-    marginBottom: 15,
-    paddingLeft: 10,
-  },
-  nameOrgInput: {
-    width: "70%",
-    height: 40,
-    borderColor: "gray",
-    borderWidth: 1,
-    marginBottom: 15,
-    paddingLeft: 10,
-  },
-  input: {
-    width: "70%",
-    height: 40,
-    borderColor: "gray",
-    borderWidth: 1,
-    marginBottom: 15,
-    paddingLeft: 10,
-  },
-  dateInput: {
-    width: "70%",
-    height: 40,
-    borderColor: "gray",
-    borderWidth: 1,
-    marginBottom: 15,
-    paddingLeft: 10,
-  },
-
-  descriptionInput: {
-    height: 80,
-  },
-  tagSelectionContainer: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    marginBottom: 15,
-  },
-  tagOption: {
-    padding: 10,
-    borderRadius: 10,
-    margin: 5,
-  },
-  card: {
-    width: 0.8 * width,
-    padding: 20,
-    backgroundColor: "#fff",
-    borderRadius: 10,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
-    marginBottom: 15,
-    alignItems: "center",
-  },
-  cardHeader: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    width: "100%",
-    marginBottom: 10,
-  },
-  cardText: {
-    fontSize: 18,
-    fontWeight: "bold",
-  },
-  cardDate: {
-    fontSize: 14,
-    color: "gray",
-  },
-  separator: {
-    height: 1,
-    width: "100%",
-    backgroundColor: "gray",
-    marginVertical: 10,
-  },
-  cardDescription: {
-    fontSize: 16,
-    textAlign: "left",
-    width: "100%",
-  },
-  tagContainer: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    marginTop: 10,
-  },
-  tag: {
-    paddingHorizontal: 10,
-    paddingVertical: 5,
-    borderRadius: 10,
-    marginRight: 5,
-    marginBottom: 5,
-  },
-  tagText: {
-    fontSize: 14,
-    color: "#fff",
-  },
-  bottomBar: {
-    width: "100%",
-    height: 60,
-    position: "absolute",
-    bottom: 0,
-    flexDirection: "row",
-    justifyContent: "space-around",
-    alignItems: "center",
-    backgroundColor: "#ddd",
-    paddingVertical: 10,
-  },
-  buttonContainer: {
-    alignItems: "center",
-  },
-  pressable: {
-    backgroundColor: '', // Background color for the pressable buttons
-    padding: 10,
-    borderRadius: 5,
-  },
-});
-``;
-
-export default styles;
